+++ src/apps/seelenweg/events.ts
@@ -1,11 +1,8 @@
 import { debounce, TimeoutIdRef } from '../Timing';
+import { toPhysicalPixels } from '../utils';
 import { PhysicalSize } from '@tauri-apps/api/dpi';
 import { emitTo, listen } from '@tauri-apps/api/event';
 import { getCurrent } from '@tauri-apps/api/webviewWindow';
-
-import { store } from './modules/shared/store/infra';
-
-import { SeelenWegMode } from '../settings/modules/seelenweg/domain';
 
 const root_container = document.getElementById('root')!;
 
@@ -17,25 +14,13 @@ export const setWindowSize = () => {
 };
 
 export const updateHitbox = () => {
-  const { margin, mode } = store.getState().settings;
-
-  root_container.style.margin = margin + 'px';
-  root_container.style.width =
-    mode === SeelenWegMode.FULL_WIDTH ? `calc(100% - ${margin * 2}px)` : 'min-content';
-
-  const width = root_container.offsetWidth;
-  const height = root_container.offsetHeight;
-
-  const windowWidth = Math.floor(width * window.devicePixelRatio);
-  const windowHeight = Math.floor(height * window.devicePixelRatio);
-
-  const screenWidth = Math.floor(window.screen.width * window.devicePixelRatio);
-  const screenHeight = Math.floor(window.screen.height * window.devicePixelRatio);
-
-  emitTo('seelenweg-hitbox', 'resize', { width: windowWidth, height: windowHeight });
+  emitTo('seelenweg-hitbox', 'resize', {
+    width: toPhysicalPixels(root_container.offsetWidth),
+    height: toPhysicalPixels(root_container.offsetHeight),
+  });
   emitTo('seelenweg-hitbox', 'move', {
-    x: Math.floor(screenWidth / 2 - windowWidth / 2),
-    y: Math.floor(screenHeight - windowHeight - Math.round(margin * window.devicePixelRatio)),
+    x: toPhysicalPixels(root_container.offsetLeft),
+    y: toPhysicalPixels(root_container.offsetTop),
   });
 };