+++ src/apps/settings/modules/seelenweg/app.ts
@@ -2,11 +2,12 @@ import { createSlice } from '@reduxjs/toolkit';
 
 import { reducersFor, selectorsFor } from '../shared/app/utils';
 
-import { SeelenWegMode, SeelenWegState } from './domain';
+import { SeelenWegMode, SeelenWegSide, SeelenWegState } from './domain';
 
 const initialState: SeelenWegState = {
   enabled: true,
   mode: SeelenWegMode.MIN_CONTENT,
+  position: SeelenWegSide.BOTTOM,
   size: 40,
   zoomSize: 70,
   margin: 8,